requirements.txt
alembic==1.7.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.1.0
matplotlib==3.5.1
multitasking==0.0.10
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandas-ta==0.3.14b0
Pillow==9.0.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
yfinance==0.1.70
zipp==3.7.0
